refactor(namer): move NAMER_LABEL to the namer leaf, off SessionService
Both PR-critic points; the second subsumes the first.

1. The declaration had been wedged between DRAFT_PR_NOTE's JSDoc and DRAFT_PR_NOTE
   itself, orphaning the doc comment above an unrelated constant and leaving
   DRAFT_PR_NOTE undocumented. My error.

2. It also lived in the CONSUMER module. Unlike DISTILL_LABEL / AUTOPILOT_LABEL /
   VERIFY_KEY_LABEL — each declared in its own producer — NAMER_LABEL sat in
   service.ts, forcing tab-reaper.ts to import the whole SessionService graph for a
   five-character string.

src/namer.ts is a zero-import leaf that service.ts already imports, and it is where
the `[a-z0-9-]` slug rule that makes the trailing space collision-proof actually
lives — so the constant is documented there, next to its own rationale. Moving it
fixes both findings at once: service.ts's JSDoc adjacency is restored, and
tab-reaper.ts now depends only on the leaf.

No behavior change: the label is still "name ". Guard re-verified — deleting the
namer boot call still turns the wiring test red.
