+++ src/index.ts
@@ -34,7 +34,7 @@ import { startTerminalTransportSelfCheck } from "./terminal-transport-metrics";
 import { generateName } from "./namer";
 import { llmName } from "./namer-llm";
 import { EventHub } from "./events";
-import { SessionService, NAMER_LABEL } from "./service";
+import { SessionService } from "./service";
 import { LearningsService } from "./learnings-service";
 import { RepoConfigService } from "./repo-config-service";
 import { StatusPoller } from "./poller";
@@ -87,6 +87,7 @@ import { StandalonePrCriticService } from "./standalone-critic";
 import { createIssueLogger } from "./issue-log";
 import { PlanGateService, shouldConsiderOnSettle } from "./plan-gate";
 import { AutopilotService, AUTOPILOT_LABEL } from "./autopilot";
+import { NAMER_LABEL } from "./namer";
 import { DrainService } from "./drain";
 import { SessionRouter, type SessionConsumer } from "./session-router";
 import { AutoMergeService } from "./automerge";
